Renamed base_collimator.py and everest_collimator.py
