+++ xcoll/beam_elements/__init__.py
@@ -1,5 +1,5 @@
-from .base_collimator import BaseCollimator
+from .base import BaseBlock, BaseCollimator
 from .absorber import BlackAbsorber
-from .everest_collimator import EverestBlock, EverestCollimator, EverestCrystal
+from .everest import EverestBlock, EverestCollimator, EverestCrystal
 
 _all_collimator_types = {BlackAbsorber, EverestCollimator, EverestCrystal}
