+++ xcoll/beam_elements/base.py
